+++ internal/oauthserver/oauth_server_test.go
@@ -18,6 +18,7 @@ import (
 	"github.com/habitat-network/habitat/internal/encrypt"
 	"github.com/habitat-network/habitat/internal/hive"
 	"github.com/habitat-network/habitat/internal/login"
+	"github.com/habitat-network/habitat/internal/core"
 	"github.com/habitat-network/habitat/internal/org"
 	"github.com/habitat-network/habitat/internal/org/testutil"
 	"github.com/habitat-network/habitat/internal/pdsclient"
@@ -715,13 +716,13 @@ func TestHandleCallbackRejectsOrgScopeForNonAdmin(t *testing.T) {
 // testIsMemberStore wraps an org.Store and overrides GetOrgByDID.
 type testIsMemberStore struct {
 	org.Store
-	fn func(ctx context.Context, did syntax.DID) (org.Org, error)
+	fn func(ctx context.Context, did syntax.DID) (core.Org, error)
 }
 
 func (s *testIsMemberStore) GetOrgForDID(
 	ctx context.Context,
 	did syntax.DID,
-) (org.Org, bool, error) {
+) (core.Org, bool, error) {
 	o, err := s.fn(ctx, did)
 	return o, false, err
 }
@@ -888,7 +889,7 @@ func TestValidate(t *testing.T) {
 	t.Run("GetOrgForDID error returns !ok with 401", func(t *testing.T) {
 		srv := newSrv(&testIsMemberStore{
 			Store: testStore(t),
-			fn: func(_ context.Context, _ syntax.DID) (org.Org, error) {
+			fn: func(_ context.Context, _ syntax.DID) (core.Org, error) {
 				return nil, errors.New("simulated database failure")
 			},
 		})
@@ -900,7 +901,7 @@ func TestValidate(t *testing.T) {
 	t.Run("non-member returns !ok with 401", func(t *testing.T) {
 		srv := newSrv(&testIsMemberStore{
 			Store: testStore(t),
-			fn: func(_ context.Context, _ syntax.DID) (org.Org, error) {
+			fn: func(_ context.Context, _ syntax.DID) (core.Org, error) {
 				return nil, org.ErrMemberNotFound
 			},
 		})
