+++ internal/org/org.go
@@ -13,7 +13,7 @@ import (
 	jose "github.com/go-jose/go-jose/v3"
 	"github.com/go-jose/go-jose/v3/jwt"
 	"github.com/habitat-network/habitat/api/habitat"
-	"github.com/habitat-network/habitat/internal/db"
+	"github.com/habitat-network/habitat/internal/core"
 	"github.com/habitat-network/habitat/internal/hive"
 	"github.com/habitat-network/habitat/internal/login"
 	"gorm.io/gorm"
@@ -42,45 +42,7 @@ func isDuplicateError(err error) bool {
 	return errors.Is(err, gorm.ErrDuplicatedKey)
 }
 
-// Org represents a single organization on a pear instance.
-type Org interface {
-	DID() syntax.DID
-	// Any app-level / further authz (like teams in an org) should happen using our permissions model.
 
-	AddAdmin(ctx context.Context, admin syntax.DID) error
-	// Only support adding members through CreateNewMemberIdentity for now
-	GetAdmins(ctx context.Context) ([]syntax.DID, error)
-	GetMembers(ctx context.Context) ([]syntax.DID, error)
-	RemoveAdmin(ctx context.Context, admin syntax.DID) error
-	RemoveMembers(ctx context.Context, members []syntax.DID) error
-	DowngradeAdmin(ctx context.Context, admin syntax.DID) error
-	IsAdmin(ctx context.Context, did syntax.DID) (bool, error)
-	IsMember(ctx context.Context, did syntax.DID) (bool, error)
-
-	// GetMetadata returns general info about this org.
-	GetMetadata(ctx context.Context, domain string) habitat.NetworkHabitatOrgGetMetadataOutput
-
-	// LoginMethod returns how users authenticate: "atproto", "google", or "password".
-	loginMethod(ctx context.Context) loginMethod
-
-	// Org member identity management; may eventually replace some of the methods above
-	IssueIdentityToken(
-		ctx context.Context,
-		caller syntax.DID,
-		reusable bool,
-		expiresAt time.Time,
-	) (token string, err error)
-	CreateNewMemberIdentity(
-		ctx context.Context,
-		token string,
-		internalHandle string,
-		password string,
-		loginID string,
-	) (*identity.Identity, error)
-	ValidateAdminSignedToken(ctx context.Context, token string) error
-
-	db.Store[Org]
-}
 
 type inviteTokenClaims struct {
 	jwt.Claims
@@ -98,7 +60,7 @@ type orgImpl struct {
 	passwordProvider *login.PasswordLoginProvider
 }
 
-var _ Org = &orgImpl{}
+var _ core.Org = &orgImpl{}
 
 // DID implements [Org].
 func (s *orgImpl) DID() syntax.DID {
@@ -354,7 +316,7 @@ func (s *orgImpl) CreateNewMemberIdentity(
 }
 
 // WithTx implements [Org].
-func (s *orgImpl) WithTx(tx *gorm.DB) Org {
+func (s *orgImpl) WithTx(tx *gorm.DB) core.Org {
 	return &orgImpl{
 		orgID:           s.orgID,
 		hive:            s.hive,
