+++ internal/org/public.go
@@ -8,6 +8,7 @@ import (
 	"github.com/bluesky-social/indigo/atproto/identity"
 	"github.com/bluesky-social/indigo/atproto/syntax"
 	"github.com/habitat-network/habitat/api/habitat"
+	"github.com/habitat-network/habitat/internal/core"
 	"gorm.io/gorm"
 )
 
@@ -120,10 +121,10 @@ func (e *everyoneOrg) AuthenticateMember(
 }
 
 // WithTx implements [Org].
-func (e *everyoneOrg) WithTx(tx *gorm.DB) Org {
+func (e *everyoneOrg) WithTx(tx *gorm.DB) core.Org {
 	return e
 }
 
-func NewEveryoneOrg() Org {
+func NewEveryoneOrg() core.Org {
 	return &everyoneOrg{}
 }
