+++ internal/org/server.go
@@ -14,6 +14,7 @@ import (
 	"github.com/gorilla/schema"
 	"github.com/habitat-network/habitat/api/habitat"
 	"github.com/habitat-network/habitat/internal/authn"
+	"github.com/habitat-network/habitat/internal/core"
 	"github.com/habitat-network/habitat/internal/instance"
 	"github.com/habitat-network/habitat/internal/pear"
 	"github.com/habitat-network/habitat/internal/permissions"
@@ -62,7 +63,7 @@ func (s *Server) IsMember(ctx context.Context, member syntax.DID) (bool, error) 
 	return true, nil
 }
 
-func (s *Server) validateOrgToken(ctx context.Context, orgID string, token string) (Org, error) {
+func (s *Server) validateOrgToken(ctx context.Context, orgID string, token string) (core.Org, error) {
 	org, err := s.store.GetOrg(ctx, syntax.DID(orgID))
 	if err != nil {
 		return nil, err
@@ -86,7 +87,7 @@ func (s *Server) GetMetadata(w http.ResponseWriter, r *http.Request) {
 	}
 
 	orgID := params.OrgId
-	var org Org
+	var org core.Org
 	// Either orgID is supplied in query params and the signed token is passed up for authn method
 	if orgID != "" {
 		org, err = s.validateOrgToken(
