+++ internal/org/store.go
@@ -10,13 +10,14 @@ import (
 
 	"github.com/bluesky-social/indigo/atproto/identity"
 	"github.com/bluesky-social/indigo/atproto/syntax"
+	"github.com/habitat-network/habitat/internal/core"
 	"github.com/habitat-network/habitat/internal/hive"
 	"github.com/habitat-network/habitat/internal/login"
 	"gorm.io/gorm"
 )
 
 type Member struct {
-	Org     Org
+	Org     core.Org
 	DID     syntax.DID
 	Role    Role
 	LoginID string
@@ -25,8 +26,8 @@ type Member struct {
 // Store is the registry of all orgs on a pear instance.
 // It routes DIDs to their org and provides cross-org membership checks.
 type Store interface {
-	GetOrg(ctx context.Context, orgID syntax.DID) (Org, error)
-	GetOrgForDID(ctx context.Context, did syntax.DID) (o Org, isMember bool, err error)
+	GetOrg(ctx context.Context, orgID syntax.DID) (core.Org, error)
+	GetOrgForDID(ctx context.Context, did syntax.DID) (o core.Org, isMember bool, err error)
 	CreateOrg(
 		ctx context.Context,
 		name string,
@@ -47,7 +48,7 @@ type storeImpl struct {
 	hive             hive.Hive
 	dir              identity.Directory
 	pearDomain       string
-	everyone         Org
+	everyone         core.Org
 	passwordProvider *login.PasswordLoginProvider
 }
 
@@ -90,7 +91,7 @@ func (s *storeImpl) orgFromModel(org *organization) (*orgImpl, error) {
 }
 
 // GetOrg returns the org with the given ID.
-func (s *storeImpl) GetOrg(ctx context.Context, orgID syntax.DID) (Org, error) {
+func (s *storeImpl) GetOrg(ctx context.Context, orgID syntax.DID) (core.Org, error) {
 	var org organization
 	if err := s.db.WithContext(ctx).Where("id = ?", orgID).First(&org).Error; err != nil {
 		return nil, ErrOrgNotFound
@@ -104,7 +105,7 @@ func (s *storeImpl) GetOrg(ctx context.Context, orgID syntax.DID) (Org, error) {
 func (s *storeImpl) GetOrgForDID(
 	ctx context.Context,
 	did syntax.DID,
-) (Org, bool /* isMember */, error) {
+) (core.Org, bool /* isMember */, error) {
 	if o, err := s.GetOrg(ctx, did); err == nil {
 		return o, false, nil
 	}
